refactor table view into data analysis component
