+++ libs/dataset-explorer/src/lib/TableView/TableViewProps.ts
@@ -3,23 +3,17 @@
 
 import { IObjectWithKey } from "@fluentui/react";
 import {
-  WeightVectorOption,
   ErrorCohort,
   JointDataset,
   ModelTypes,
   ITelemetryEvent
 } from "@responsible-ai/core-ui";
 
-export interface IIndividualFeatureImportanceProps {
-  allSelectedItems: IObjectWithKey[];
+export interface ITableViewProps {
   features: string[];
   jointDataset: JointDataset;
-  invokeModel?: (data: any[], abortSignal: AbortSignal) => Promise<any[]>;
-  selectedWeightVector: WeightVectorOption;
-  weightOptions: WeightVectorOption[];
-  weightLabels: any;
-  onWeightChange: (option: WeightVectorOption) => void;
   selectedCohort: ErrorCohort;
   modelType?: ModelTypes;
+  onAllSelectedItemsChange: (allSelectedItems: IObjectWithKey[]) => void;
   telemetryHook?: (message: ITelemetryEvent) => void;
 }
